LogOut/requirements.txt
requests~=2.31.0
certifi           ~= 2024.2.2
charset-normalizer ~=3.3.2
grpcio             ~=1.62.1
grpcio-tools       ~=1.62.1
idna               ~=3.6
pip                ~=23.2.1
protobuf           ~=4.25.3
requests           ~=2.31.0
setuptools         ~=69.1.1
urllib3            ~=2.2.1


